requirements.txt
python-telegram-bot>=20.8,<21.0
python-dotenv>=1.0.0
requests>=2.31.0
python-decouple>=3.8
aiosqlite>=0.19.0
qrcode>=7.4.2
Pillow>=10.0.0
